Profile Display Update: Changed to /uid + Stat Highlighting
Now Profiles can load using:
/starrail/profile/####UID####!

Highlighting behavior applied to all of:
- Skill Tree Nodes
- Stat Section
- Light Cone Stats
- Relic Main Stat
- Relic Sub Stats
- Relic Set effect
